+++ lib/Map/ImageryProvider/ProtomapsImageryProvider.ts
@@ -5,36 +5,35 @@ import circle from "@turf/circle";
 import { Feature } from "@turf/helpers";
 import i18next from "i18next";
 import { cloneDeep, isEmpty } from "lodash-es";
-import { action, observable, runInAction, makeObservable } from "mobx";
+import { action, makeObservable, observable, runInAction } from "mobx";
 import {
   Bbox,
-  Feature as ProtomapsFeature,
   GeomType,
-  Labelers,
   LabelRule,
+  Labelers,
   LineSymbolizer,
-  painter,
+  Rule as PaintRule,
   PmtilesSource,
   PreparedTile,
-  Rule as PaintRule,
+  Feature as ProtomapsFeature,
   TileCache,
   TileSource,
   View,
   Zxy,
-  ZxySource
+  ZxySource,
+  painter
 } from "protomaps";
 import Cartographic from "terriajs-cesium/Source/Core/Cartographic";
 import Credit from "terriajs-cesium/Source/Core/Credit";
-import defaultValue from "terriajs-cesium/Source/Core/defaultValue";
 import DeveloperError from "terriajs-cesium/Source/Core/DeveloperError";
 import CesiumEvent from "terriajs-cesium/Source/Core/Event";
 import CesiumMath from "terriajs-cesium/Source/Core/Math";
 import Rectangle from "terriajs-cesium/Source/Core/Rectangle";
 import WebMercatorTilingScheme from "terriajs-cesium/Source/Core/WebMercatorTilingScheme";
+import defaultValue from "terriajs-cesium/Source/Core/defaultValue";
 import ImageryLayerFeatureInfo from "terriajs-cesium/Source/Scene/ImageryLayerFeatureInfo";
-import filterOutUndefined from "../../Core/filterOutUndefined";
-import isDefined from "../../Core/isDefined";
 import TerriaError from "../../Core/TerriaError";
+import isDefined from "../../Core/isDefined";
 import {
   FeatureCollectionWithCrs,
   FEATURE_ID_PROP as GEOJSON_FEATURE_ID_PROP,
@@ -85,6 +84,8 @@ export type ProtomapsData = string | FeatureCollectionWithCrs | Source;
 interface Options {
   terria: Terria;
 
+  /** This must be defined to support pickedFeatures in share links */
+  id?: string;
   data: ProtomapsData;
   minimumZoom?: number;
   maximumZoom?: number;
@@ -272,6 +273,10 @@ export default class ProtomapsImageryProvider
   readonly errorEvent = new CesiumEvent();
   readonly ready = true;
   readonly credit: Credit;
+  /** This is only used for Terria feature picking - as we track ImageryProvider feature picking by url (See PickedFeatures/Cesium._attachProviderCoordHooks). This URL is never called.
+   * This is set using the `id` property in the constructor options
+   */
+  readonly url?: string;
 
   // Set values to please poor cesium types
   readonly defaultNightAlpha = undefined;
@@ -349,6 +354,7 @@ export default class ProtomapsImageryProvider
     }
 
     this.errorEvent = new CesiumEvent();
+    this.url = options.id;
 
     this.ready = true;
 
@@ -544,10 +550,12 @@ export default class ProtomapsImageryProvider
       this.source instanceof GeojsonSource &&
       this.source.geojsonObject
     ) {
+      // Get rough meters per pixel (at equator) for given zoom level
+      const zoomMeters = 156543 / Math.pow(2, level);
       // Create circle with 10 pixel radius to pick features
       const buffer = circle(
         [CesiumMath.toDegrees(longitude), CesiumMath.toDegrees(latitude)],
-        10 * this.terria.mainViewer.scale,
+        10 * zoomMeters,
         {
           steps: 10,
           units: "meters"
@@ -662,6 +670,7 @@ export default class ProtomapsImageryProvider
 
     return new ProtomapsImageryProvider({
       terria: options?.terria ?? this.terria,
+      id: options?.id ?? this.url,
       data,
       minimumZoom: options?.minimumZoom ?? this.minimumLevel,
       maximumZoom: options?.maximumZoom ?? this.maximumLevel,