+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Fix `GeoJsonMixin` timeline animation for lines/polygons
 - Fix bug in mismatched GeoJSON Feature `_id_` and TableMixin `rowId` - this was causing incorrect styling when using `filterByProperties` or features had `null` geometry
 - Fix splitter for `GeoJsonMixin` (lines and polygon features only)
+- Fix share links with picked features from `ProtomapsImageryProvider`
 - [The next improvement]
 
 #### 8.3.6 - 2023-10-03